# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API DOCS",
    "description": "Api document for file-browser-backend",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/download": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "file_path",
            "required": true,
            "schema": "string",
            "description": "path to file"
          },
          {
            "in": "query",
            "name": "s3",
            "schema": "string",
            "description": "using s3 or not"
          },
          {
            "in": "query",
            "name": "bucket",
            "schema": "string",
            "description": "if use s3, this is bucket use"
          }
        ],
        "summary": "Download file with query string is path",
        "responses": {
          "400": {
            "description": "Error when the file is not existed",
            "schema": {
              "type": "object",
              "properties": {
                "messsage": {
                  "type": "string",
                  "description": "error message"
                }
              }
            }
          }
        }
      }
    },
    "/file-explorer/recursive": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "dir",
            "required": true,
            "schema": "string",
            "description": "path that you want to explore"
          },
          {
            "in": "query",
            "name": "s3",
            "schema": "string",
            "description": "if you want to use s3"
          },
          {
            "in": "query",
            "name": "bucket",
            "schema": "string",
            "description": "bucket you want to explore"
          }
        ],
        "summary": "Explorer everything inside",
        "responses": {
          "400": {
            "description": "Error",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "error message"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "description": "data"
                }
              }
            }
          }
        }
      }
    },
    "/file-explorer/shallow": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "dir",
            "required": true,
            "schema": "string",
            "description": "path that you want to explore"
          },
          {
            "in": "query",
            "name": "s3",
            "schema": "string",
            "description": "if you want to use s3"
          },
          {
            "in": "query",
            "name": "bucket",
            "schema": "string",
            "description": "bucket you want to explore"
          }
        ],
        "summary": "Explorer 1 level inside",
        "responses": {
          "400": {
            "description": "Error",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "error message"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "description": "data"
                }
              }
            }
          }
        }
      }
    },
    "/upload": {
      "post": {
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "upload-file",
            "schema": "string",
            "description": "field that need for multipart/form-data",
            "required": true
          },
          {
            "in": "query",
            "name": "location",
            "required": true,
            "schema": "string",
            "description": "folder to store file"
          },
          {
            "in": "query",
            "name": "s3",
            "schema": "string",
            "description": "using s3 or not"
          },
          {
            "in": "query",
            "name": "bucket",
            "schema": "string",
            "description": "if use s3, this is bucket use"
          }
        ],
        "summary": "Upload a file",
        "responses": {
          "400": {
            "description": "Error when the file is not existed",
            "schema": {
              "type": "object",
              "properties": {
                "messsage": {
                  "type": "string",
                  "description": "error message"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "description": "data object"
                }
              }
            }
          }
        }
      }
    }
  }
}